Add gzip wrappers for iostreams
Add a tiny library with two functions `gzip_istream` and `gzip_ostream`
which wrap an i/ostream with a boost filtering stream, adding gzip
(de-)compression.

The istream wrapper first reads the first two bytes of the input to
determine if the input is really gzip-compressed.  If not, no
decompression is added (i.e. the wrapper does nothing).  I'm not sure if
this behaviour (i.e. reading the first bytes, then resetting the
pointer) might cause issues in some cases.  At least for fstreams it
seems to be fine.

The ostream wrapper has an optional argument `compress`.  If set to
false, no compression is added by the wrapper (i.e. it basically does
nothing).  This can be useful to keep the code clean in a setup where
compression can be enabled/disabled by a parameter.